+++ src/scopes/internal/zmq_middleware/StopPublisher.cpp
@@ -21,7 +21,6 @@
 #include <unity/scopes/ScopeExceptions.h>
 
 #include <cassert>
-#include <iostream>  // TODO: remove this once logging is added
 
 using namespace std;