+++ test/gtest/scopes/internal/smartscopes/SmartScopesClient/SmartScopesClient_test.cpp
@@ -16,12 +16,16 @@
  * Authored by: Marcus Tomlinson <marcus.tomlinson@canonical.com>
  */
 
-#include <unity/scopes/OptionSelectorFilter.h>
-#include <unity/scopes/internal/smartscopes/HttpClientQt.h>
 #include <unity/scopes/internal/JsonCppNode.h>
+#include <unity/scopes/internal/Logger.h>
+#include <unity/scopes/internal/smartscopes/HttpClientQt.h>
 #include <unity/scopes/internal/smartscopes/SmartScopesClient.h>
+#include <unity/scopes/OptionSelectorFilter.h>
 
 #include <unity/UnityExceptions.h>
+#include <unity/util/FileIO.h>
+#include <boost/filesystem/operations.hpp>
+#include <fstream>
 
 #include "../RaiiServer.h"
 
@@ -35,6 +39,8 @@ using namespace unity::scopes::internal;
 using namespace unity::scopes::internal::smartscopes;
 using namespace unity::test::scopes::internal::smartscopes;
 
+BOOST_LOG_INLINE_GLOBAL_LOGGER_DEFAULT(test_logger, boost::log::sources::severity_channel_logger_mt<>)
+
 namespace
 {
 
@@ -44,10 +50,26 @@ public:
     SmartScopesClientTest()
         : http_client_(new HttpClientQt(20000)),
           json_node_(new JsonCppNode()),
-          server_(FAKE_SSS_PATH)
+          server_(FAKE_SSS_PATH, FAKE_SSS_LOG)
     {
+        boost::filesystem::remove(FAKE_SSS_LOG);
         sss_url_ = "http://127.0.0.1:" + std::to_string(server_.port_);
-        ssc_ = std::make_shared<SmartScopesClient>(http_client_, json_node_, sss_url_);
+        ssc_ = std::make_shared<SmartScopesClient>(http_client_, json_node_, test_logger::get(), sss_url_, PARTNER_FILE);
+    }
+
+    bool grep_string(std::string const &s)
+    {
+        std::stringstream str(unity::util::read_text_file(FAKE_SSS_LOG));
+        while (str)
+        {
+            char tmp[1024];
+            str.getline(tmp, 1024);
+            if (strstr(tmp, s.c_str()))
+            {
+                return true;
+            }
+        }
+        return false;
     }
 
 protected:
@@ -85,7 +107,7 @@ TEST_F(SmartScopesClientTest, remote_scopes)
     EXPECT_EQ(0, scopes[0].version);
     EXPECT_EQ(nullptr, scopes[0].appearance);
     EXPECT_EQ(nullptr, scopes[0].settings);
-    EXPECT_TRUE(scopes[0].tags.empty());
+    EXPECT_TRUE(scopes[0].keywords.empty());
 
     EXPECT_EQ("dummy.scope.2", scopes[1].id);
     EXPECT_EQ("Dummy Demo Scope 2", scopes[1].name);
@@ -99,7 +121,7 @@ TEST_F(SmartScopesClientTest, remote_scopes)
     EXPECT_EQ("#00BEEF", (*scopes[1].appearance)["background"].get_string());
     EXPECT_EQ("logo.png", (*scopes[1].appearance)["PageHeader"].get_dict()["logo"].get_string());
     EXPECT_EQ(nullptr, scopes[1].settings);
-    EXPECT_TRUE(scopes[1].tags.empty());
+    EXPECT_TRUE(scopes[1].keywords.empty());
 
     EXPECT_EQ("dummy.scope.3", scopes[2].id);
     EXPECT_EQ("Dummy Demo Scope 3", scopes[2].name);
@@ -117,11 +139,21 @@ TEST_F(SmartScopesClientTest, remote_scopes)
               "\"defaultValue\":23},\"type\":\"number\"},{\"displayName\":\"Enabled\",\"id\":"
               "\"enabled\",\"parameters\":{\"defaultValue\":true},\"type\":\"boolean\"}]\n",
               *scopes[2].settings);
-    ASSERT_EQ(4, scopes[2].tags.size());
-    EXPECT_EQ("music", scopes[2].tags[0]);
-    EXPECT_EQ("video", scopes[2].tags[1]);
-    EXPECT_EQ("news", scopes[2].tags[2]);
-    EXPECT_EQ("games", scopes[2].tags[3]);
+    ASSERT_EQ(4, scopes[2].keywords.size());
+    EXPECT_EQ("music", scopes[2].keywords[0]);
+    EXPECT_EQ("video", scopes[2].keywords[1]);
+    EXPECT_EQ("news", scopes[2].keywords[2]);
+    EXPECT_EQ("games", scopes[2].keywords[3]);
+
+    EXPECT_TRUE(grep_string("/remote-scopes : partner_id=Partner%20String"));
+}
+
+TEST_F(SmartScopesClientTest, remote_scopes_no_partner)
+{
+    std::vector<RemoteScope> scopes;
+    auto ssc_no_partner_ = std::make_shared<SmartScopesClient>(http_client_, json_node_, test_logger::get(), sss_url_, "/this/file/doesnt/exist");
+    EXPECT_TRUE(ssc_no_partner_->get_remote_scopes(scopes, "", false));
+    EXPECT_FALSE(grep_string("/remote-scopes : partner_id"));
 }
 
 TEST_F(SmartScopesClientTest, search)